% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\name{Events}
\alias{Events}
\title{Events}
\usage{
Events(events_path, crs)
}
\arguments{
\item{events_path}{Select path of events txt file.}

\item{crs}{Coordinate system used. Strongly recommended the use of EPSG.}
}
\description{
Reads and uploads txt events file.
}
