% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.R
\name{Line}
\alias{Line}
\title{Line}
\usage{
Line(line_path, crs)
}
\arguments{
\item{line_path}{Line used in Siriema - txt file.}

\item{crs}{Coordinate system used. Preferably EPSG.}
}
\description{
Reads and uploads txt road file.
}
