% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Milepost.R
\name{Milepost}
\alias{Milepost}
\title{Milepost}
\usage{
Milepost(road, distance, dist_or_n = "dist")
}
\arguments{
\item{road}{Road shapefile.}

\item{distance}{Distance of stakes.}

\item{dist_or_n}{Selects distance of stakes or number of stakes}
}
\description{
Calculates two number of classes of Fisher-Jenks algorithm
}
