% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siriemashapes.R
\name{siriemashapes}
\alias{siriemashapes}
\title{siriemashapes}
\usage{
siriemashapes(line_path, crs, events_path, hotspot_path)
}
\arguments{
\item{line_path}{Line used in Siriema - txt file.}

\item{crs}{Coordinate system used. Preferably EPSG.}

\item{events_path}{Events used in Siriema - txt file.}

\item{hotspot_path}{Hotspot results obtained from Siriema - txt or dat file.}
}
\description{
Creates complete dataframe/shapefile from Hotspots results obtained from Siriema software
}
