% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siriemashapes.R
\name{siriemashapes}
\alias{siriemashapes}
\title{siriemashapes}
\usage{
siriemashapes(line_path, events_path, hotspot_path, crs)
}
\arguments{
\item{line_path}{Line used in Siriema - txt file path.}

\item{events_path}{Events used in Siriema - txt file path.}

\item{hotspot_path}{Hotspot results obtained from Siriema - txt or dat file path.}

\item{crs}{Coordinate reference system. Strongly recommended the use of EPSG number.}
}
\description{
Creates complete dataframe/shapefile from Hotspots results obtained from \href{http://www.ufrgs.br/siriema/}{Siriema} software.
}
\details{
\strong{WARNING}: For best (and accurate) results, it is strongly recommended that the Hotspot analysis is performed taking into account the number of divisions set to double of the radius.

The function offers a large list object containing two elements: a spatial data frame with improvements of Siriema's Hotspot Analysis results and a summary of the species found in each sample unit.

The spatial data frame is called \code{shapefile} and it is composed by the following attributes:
\itemize{
\item \strong{ID}: ID of the sample unit.
\item \strong{length}: Length of the sample unit. It is supposed to be very similar to the diameter (double of the radius) of the original Hotspot Analysis.
\item \strong{km_center}: Central km mark of the sample unit.
\item \strong{km_ini}: Initial km mark of the sample unit.
\item \strong{km_end}: Final km mark of the sample unit.
\item \strong{HS}: Aggregation intensity.
\item \strong{UCL}: Upper confidence limit.
\item \strong{LCL}: Lower confidence limit.
\item \strong{HS-UCL}: Subtraction of aggregation intensity by the upper limit of confidence interval. This is the "real" value of aggregation intensity as it already takes into account what is above the confidence interval.
\item \strong{Hotspot}: Information if the sample unit is a hotspot. Y means that \eqn{HS-UCL > 0} and therefore it is already a hotspot.
\item \strong{NEvents}: Number of events in the sample unit.
\item \strong{Rank}: Rank of aggregation intensity based on \emph{HS-UCL}.
\item \strong{FisherJenks3}: Fisher-Jenks classification in three classes (text) based on \emph{HS-UCL}. Calculation is performed only on rows that already are a hotspot.
\item \strong{FJ3num}: Fisher-Jenks classification in three classes (numeric) based on \emph{HS-UCL}. Calculation is performed only on rows that already are a hotspot.
\item \strong{FisherJenks5}: Fisher-Jenks classification in five classes (text) based on \emph{HS-UCL}. Calculation is performed only on rows that already are a hotspot.
\item \strong{FJ5num}: Fisher-Jenks classification in five classes (numeric) based on \emph{HS-UCL}. Calculation is performed only on rows that already are a hotspot.
}

The summary of the species found in each sample unit is called \code{species_df} and it is composed by the following attributes:
\itemize{
\item \strong{ID}: ID of the sample unit.
\item \strong{Sp}: Species name (as from the Events file)
\item \strong{NEvents}: Number of events in the sample unit.
}
}
\author{
Rubem Dornas \email{(rapdornas@gmail.com)}
}
